refactoring meta all - remove peromyscus
While refactoring the meta analysis on all data I found that the
Peromyscus dataset was breaking the meta-regression.  The reasons are
not totally clear, except that it was already obvious that Peromyscus
was weird from previous graphs, notably having only very low dosages,
not sufficient to fit the curves derived.  I removed it from the
original filter and tidied up a few more things along the way.